Splits FHIR storage test to account for differing R4 and STU3 behaviour.
